Feature/api permissions

See merge request BI-TWA/B231/team-pendrej!14

+++ src/Api/Controller/ReservationController.php
@@ -2,12 +2,15 @@
 
 namespace App\Api\Controller;
 
+use App\Api\Model\AppUserInput;
 use App\Api\Model\ReservationInput;
 use App\Api\Model\ReservationOutput;
+use App\Entity\AppUser;
 use App\Entity\Reservation;
 use App\Service\AppUserManager;
 use App\Service\ReservationManager;
 use App\Service\RoomManager;
+use App\Voter\ReservationVoter;
 use Exception;
 use FOS\RestBundle\Controller\AbstractFOSRestController;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
@@ -29,6 +32,7 @@ class ReservationController extends AbstractFOSRestController
     #[Rest\View(statusCode: 200)]
     public function list(Request $request): array
     {
+        $this->denyAccessUnlessGranted(ReservationVoter::VIEW_INDEX_ALL);
         $title = $request->query->get('title');
         $description = $request->query->get('description');
 
@@ -38,6 +42,7 @@ class ReservationController extends AbstractFOSRestController
                 $this->generateUrlIfNotNull($entity->getRoom(), 'api_rooms_detail'),
                 $this->generateUrlIfNotNull($entity->getApprovedBy(), 'api_app_users_detail'),
                 $this->generateUrlIfNotNull($entity->getReservedFor(), 'api_app_users_detail'),
+                $this->getVisitorsUrls($entity),
             ),
             $this->reservationManager->findReservationsByFilters($title, $description)
         );
@@ -50,6 +55,7 @@ class ReservationController extends AbstractFOSRestController
     public function detail(int $id): ReservationOutput
     {
         $reservation = $this->reservationManager->findById($id);
+        $this->denyAccessUnlessGranted(ReservationVoter::VIEW_DETAIL, $reservation);
         if (!$reservation) {
             throw new HttpException(404, 'Reservation not found');
         }
@@ -59,6 +65,7 @@ class ReservationController extends AbstractFOSRestController
             $this->generateUrlIfNotNull($reservation->getRoom(), 'api_rooms_detail'),
             $this->generateUrlIfNotNull($reservation->getApprovedBy(), 'api_app_users_detail'),
             $this->generateUrlIfNotNull($reservation->getReservedFor(), 'api_app_users_detail'),
+            $this->getVisitorsUrls($reservation),
         );
     }
 
@@ -71,7 +78,13 @@ class ReservationController extends AbstractFOSRestController
     #[Rest\View(statusCode: 201)]
     public function update(?int $id, ReservationInput $reservationInput, ConstraintViolationListInterface $errors): ReservationOutput
     {
-        $reservation = $id !== null ? $this->findOrFail($id) : new Reservation();
+        if ($id !== null) {
+            $reservation = $this->findOrFail($id);
+            $this->denyAccessUnlessGranted(ReservationVoter::EDIT, $reservation);
+        } else {
+            $reservation = new Reservation();
+            $this->denyAccessUnlessGranted(ReservationVoter::CREATE);
+        }
 
         if ($errors->count() > 0) {
             throw new HttpException(400, message: \implode("\n", \array_map(
@@ -88,6 +101,7 @@ class ReservationController extends AbstractFOSRestController
             $this->generateUrlIfNotNull($reservation->getRoom(), 'api_rooms_detail'),
             $this->generateUrlIfNotNull($reservation->getApprovedBy(), 'api_app_users_detail'),
             $this->generateUrlIfNotNull($reservation->getReservedFor(), 'api_app_users_detail'),
+            $this->getVisitorsUrls($reservation),
         );
     }
 
@@ -95,6 +109,7 @@ class ReservationController extends AbstractFOSRestController
     #[Rest\View(statusCode: 204)]
     public function delete(int $id): void
     {
+        $this->denyAccessUnlessGranted(ReservationVoter::DELETE);
         $reservation = $this->findOrFail($id);
         $this->reservationManager->deleteFromDatabase($reservation);
     }
@@ -107,8 +122,10 @@ class ReservationController extends AbstractFOSRestController
     public function approve(int $id): ReservationOutput
     {
         $reservation = $this->findOrFail($id);
+        $this->denyAccessUnlessGranted(ReservationVoter::CAN_APPROVE, $reservation);
 
         $reservation->setStatus('approved');
+        $reservation->setApprovedBy($this->getUser());
         $reservation = $this->reservationManager->saveToDatabase($reservation);
 
         return ReservationOutput::fromEntity(
@@ -116,6 +133,7 @@ class ReservationController extends AbstractFOSRestController
             $this->generateUrlIfNotNull($reservation->getRoom(), 'api_rooms_detail'),
             $this->generateUrlIfNotNull($reservation->getApprovedBy(), 'api_app_users_detail'),
             $this->generateUrlIfNotNull($reservation->getReservedFor(), 'api_app_users_detail'),
+            $this->getVisitorsUrls($reservation),
         );
     }
 
@@ -127,6 +145,7 @@ class ReservationController extends AbstractFOSRestController
     public function reject(int $id): ReservationOutput
     {
         $reservation = $this->findOrFail($id);
+        $this->denyAccessUnlessGranted(ReservationVoter::CAN_REJECT, $reservation);
 
         $reservation->setStatus('rejected');
         $reservation = $this->reservationManager->saveToDatabase($reservation);
@@ -136,8 +155,40 @@ class ReservationController extends AbstractFOSRestController
             $this->generateUrlIfNotNull($reservation->getRoom(), 'api_rooms_detail'),
             $this->generateUrlIfNotNull($reservation->getApprovedBy(), 'api_app_users_detail'),
             $this->generateUrlIfNotNull($reservation->getReservedFor(), 'api_app_users_detail'),
+            $this->getVisitorsUrls($reservation),
         );
     }
+
+    #[Rest\Patch('/reservation/{id}/visitor', name: 'api_reservations_add_visitor', requirements: ['id' => '\d+'])]
+    #[ParamConverter('appUserInput', converter: 'fos_rest.request_body')]
+    #[Rest\View(statusCode: 200)]
+    public function addVisitor(int $id, AppUserInput $appUserInput): ReservationOutput
+    {
+        $reservation = $this->findOrFail($id);
+        $user = $this->findOrFailUser($appUserInput->id);
+
+        $reservation->addVisitor($user);
+        $reservation = $this->reservationManager->saveToDatabase($reservation);
+        return ReservationOutput::fromEntity(
+            $reservation,
+            $this->generateUrlIfNotNull($reservation->getRoom(), 'api_rooms_detail'),
+            $this->generateUrlIfNotNull($reservation->getApprovedBy(), 'api_app_users_detail'),
+            $this->generateUrlIfNotNull($reservation->getReservedFor(), 'api_app_users_detail'),
+            $this->getVisitorsUrls($reservation),
+        );
+    }
+
+    #[Rest\Delete('/reservation/{id}/visitor/{visitorId}', name: 'api_reservations_remove_visitor', requirements: ['id' => '\d+', 'visitorId' => '\d+'])]
+    #[Rest\View(statusCode: 204)]
+    public function removeVisitor(int $id, int $visitorId): void
+    {
+        $reservation = $this->findOrFail($id);
+        $user = $this->findOrFailUser($visitorId);
+
+        $reservation->removeVisitor($user);
+        $this->reservationManager->saveToDatabase($reservation);
+    }
+
 
     private function findOrFail(int $id): Reservation
     {
@@ -149,10 +200,29 @@ class ReservationController extends AbstractFOSRestController
         return $reservation;
     }
 
+    private function findOrFailUser(int $id): AppUser
+    {
+        $user = $this->appUserManager->userRepository->find($id);
+
+        if (!$user) {
+            throw $this->createNotFoundException('User not found');
+        }
+
+        return $user;
+    }
+
     private function generateUrlIfNotNull(?object $entity, string $routeName): ?string {
         if ($entity !== null && method_exists($entity, 'getId')) {
             return $this->generateUrl($routeName, ['id' => $entity->getId()]);
         }
         return null;
+    }
+
+    private function getVisitorsUrls(Reservation $reservation): array
+    {
+        return array_map(
+            fn (AppUser $entity) => $this->generateUrl('api_app_users_detail', ['id' => $entity->getId()]),
+            $reservation->getVisitors()->toArray()
+        );
     }
 }
