+++ src/Api/Controller/RoomController.php
@@ -4,12 +4,15 @@ namespace App\Api\Controller;
 
 use App\Api\Model\RoomInput;
 use App\Api\Model\RoomOutput;
+use App\Entity\AppUser;
+use App\Entity\Reservation;
 use App\Entity\Room;
 use App\Repository\BuildingRepository;
 use App\Repository\GroupRepository;
 use App\Service\AppUserManager;
 use App\Service\GroupManager;
 use App\Service\RoomManager;
+use App\Voter\RoomVoter;
 use Exception;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Component\HttpFoundation\Request;
@@ -33,12 +36,18 @@ class RoomController extends AbstractFOSRestController
 
     public function list(Request $request): array
     {
+        $this->denyAccessUnlessGranted(RoomVoter::VIEW_INDEX);
         $name = $request->query->get('name');
         $code = $request->query->get('code');
         $buildingId = $request->query->get('buildingId');
 
         $rooms = array_map(
-            fn (Room $entity) => RoomOutput::fromEntity($entity, $this->getUsersUrls($entity, true), $this->getUsersUrls($entity, false), $this->getGroupsUrls($entity)),
+            fn (Room $entity) => RoomOutput::fromEntity(
+                $entity,
+                $this->getUsersUrls($entity, true),
+                $this->getUsersUrls($entity, false),
+                $this->getGroupsUrls($entity),
+                $this->getReservationsUrls($entity)),
             $this->roomManager->findRoomsByFilters($name, $code, $buildingId)
         );
 
@@ -50,12 +59,19 @@ class RoomController extends AbstractFOSRestController
     public function detail(int $id): RoomOutput
     {
         $room = $this->roomManager->getRoomById($id);
+        $this->denyAccessUnlessGranted(RoomVoter::VIEW_DETAIL, $room);
 
         if (!$room) {
             throw $this->createNotFoundException('Room not found');
         }
 
-        return RoomOutput::fromEntity($room, $this->getUsersUrls($room, true), $this->getUsersUrls($room, false), $this->getGroupsUrls($room));
+        return RoomOutput::fromEntity(
+            $room,
+            $this->getUsersUrls($room, true),
+            $this->getUsersUrls($room, false),
+            $this->getGroupsUrls($room),
+            $this->getReservationsUrls($room)
+        );
     }
 
     /**
@@ -67,7 +83,13 @@ class RoomController extends AbstractFOSRestController
     #[Rest\View(statusCode: 201)]
     public function update(?int $id, RoomInput $roomInput, ConstraintViolationListInterface $errors): RoomOutput
     {
-        $room = $id !== null ? $this->findOrFail($id) : new Room();
+        if ($id !== null) {
+            $room = $this->findOrFail($id);
+            $this->denyAccessUnlessGranted(RoomVoter::EDIT, $room);
+        } else {
+            $this->denyAccessUnlessGranted(RoomVoter::CREATE);
+            $room = new Room();
+        }
 
         if ($errors->count() > 0) {
             throw new HttpException(400, message: \implode("\n", \array_map(
@@ -82,18 +104,38 @@ class RoomController extends AbstractFOSRestController
             $room,
             $this->getUsersUrls($room, true),
             $this->getUsersUrls($room, false),
-            $this->getGroupsUrls($room));
+            $this->getGroupsUrls($room),
+            $this->getReservationsUrls($room)
+        );
     }
 
     #[Rest\Delete('/room/{id}', name: 'api_rooms_delete', requirements: ['id' => '\d+'])]
     #[Rest\View(statusCode: 204)]
     public function delete(int $id): void
     {
+        $this->denyAccessUnlessGranted(RoomVoter::DELETE);
         $room = $this->findOrFail($id);
         $this->roomManager->deleteFromDatabase($room);
     }
 
-    public function findOrFail(int $id): Room
+    // ověření zda uživatel má v tuto chvíli přístup do místnosti (je neobsazená, je jejím uživatelem, má schválenou rezervaci)
+    #[Rest\Get('/room/{id}/access', name: 'api_rooms_access', requirements: ['id' => '\d+'])]
+    #[Rest\View(statusCode: 200)]
+    public function hasAccess(int $id): bool
+    {
+        $room = $this->findOrFail($id);
+        /** @var AppUser $user */
+        $user = $this->getUser();
+        if ($this->isGranted(RoomVoter::HAS_FULL_ACCESS_TO_ROOM, $room)
+            || $this->roomManager->isRoomFree($room)
+            || $this->roomManager->hasApprovedReservation($room, $user)
+        ) {
+            return true;
+        }
+        return false;
+    }
+
+    private function findOrFail(int $id): Room
     {
         $room = $this->roomManager->getRoomById($id);
         if (!$room) {
@@ -102,7 +144,7 @@ class RoomController extends AbstractFOSRestController
         return $room;
     }
 
-    public function getUsersUrls(Room $room, bool $isMember): array
+    private function getUsersUrls(Room $room, bool $isMember): array
     {
         if ($isMember) {
             return array_map(
@@ -118,11 +160,19 @@ class RoomController extends AbstractFOSRestController
         }
     }
 
-    public function getGroupsUrls(Room $room): array
+    private function getGroupsUrls(Room $room): array
     {
         return array_map(
             fn ($group) => $this->generateUrl('api_groups_detail', ['id' => $group->getId()]),
             $room->getOwningGroups()->toArray()
+        );
+    }
+
+    private function getReservationsUrls(Room $room): array
+    {
+        return array_map(
+            fn ($reservation) => $this->generateUrl('api_reservations_detail', ['id' => $reservation->getId()]),
+            $room->getReservations()->toArray()
         );
     }
 }
